requirements/common.txt
asgiref==3.9.1
coverage==7.10.5
Django==5.2.5
django-crum==0.7.9
django-environ==0.12.0
djangorestframework==3.16.1
pillow==11.3.0
pygraphviz==1.14
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
supervisor==4.3.0
tzdata==2025.2
uWSGI==2.0.30
